Returned objects should be java.io.Serializable
Objects returned by calls to CloudFoundryClient ought to be Serializable
to allow keeping state in apps that can only rely on this (e.g. replicated
HttpSession or in my case, Android Bundle).
There should be no harm in allowing this.

Went with generated uid, as this is what was done for the Exception there already.

Change-Id: I168cc53fe0d03a1f7d3ac37fbc639e39dfe99652