Add space user role assignation for current user.
As space creation does not handle user role, it has to be done after. So
once a space is created, current logged user can be assign to it.
Assigned user roles are identical to CF CLI ones, ie Manager, Developer
and Auditor.

Only assign and list are implemented.

Related to PR #238