Update API to use Reactive Streams
Previously, the API exposed to users depended on RxJava.  While this
was a reasonable choice, the maintainer of RxJava has stated that API
should exposed Reactive Streams types instead.  This change updates
the API to do this.  In addition to the API change, the Spring
implementation of the API and the integration tests were updated to
use the Reactor project.