Test Structure Fixes
Previously all of the tests in the Spring implementation used the
consume() method to perform their assertions.  Because consume()
swallows exceptions, test failures were not propagated.  This change
updates the tests so that they propagate exceptions.  Only two tests
were found to be failing (and hidden) and both were bugs in the test
code, not the main code.