First Integration Test
This change adds the first "real" integration test to the project.  At
the moment the test in this change does not pass.  This is to be used
as a way to create the rest of the integration test infrastructure.
When this test passes, the initial task of integration testing will be
complete and ready for more tests to be written.