Test mockRequest refactoring
Previously mock requests were configured using an overloaded method
using a variable number of strings.  This led to confusion an overall
lack of readability for this configuration.  This changes the
configuration to use a "builder" type.  This leaves the configuration
with the readability of using mockMvc directly, while removing the
boilerplate required for our tests.