Project Lombok
Previously, the vast majority of the code in project was dedicated to
getters and setters along with their tests.  Since this code is quite
boilerplate, removing that code would be a big win in terms of the
time taken to deliver a given API.

This change introduces Project Lombok which automatically writes
getters, setters, and builders for each type.  While it requires an
IDE plugin in order to properly view the code, I consider this a win
compared to the amount of code that used to exist.