cf:app fails when there is no usage information.
Previously running cf:app for an application that has not exited staging
results in a NPE in the rendering code. This commit fixes this problem by
omitting the relevant information from the rendered output in this case.

Test were added to check that rendering copes with and without usage information.

[#105493832]