Platform as Parent
Previously the Spring IO platform was included in the project as an
imported dependency.  While this enabled us to take advantage of the
dependency management in the platform, it wasn't the best way to use
it.  This change makes the platform the parent of the project instead
which allows us to override properties rather than overriding both
properties and dependencyManagement declarations.