List all Developers for the Space (v2)
This change adds the v2 API to List all Developers for the Space
(GET /v2/spaces/:guid/developers).

This commit pulls out ListSpaceMembersResource which is a common
response resource for multiple responses. This new resource is
consumed in a similar way to the v2/applications code. The common
SpaceResource and SpaceEntity has also been refactored to match the
style used in the v2/applications code.

[#101419374]