List all Service Instances for the Space (v2)
This change adds the v2 API to List all Service Instances for the Space
(GET /v2/spaces/:guid/service_instances).

This commit also refactors the v2/serviceInstance code to match the
style used in the rest of the v2 code.

This commit also changes the SecurityGroupEntity to use a List return
type instead of an Array for better Lombok compatability.

[#101419384]