Fix NPE when parsing application State
Instead of falling over with an NPE when the date value can't be
parsed then Client API will simple return an Application Usage state
with a null value for the Date of the stats. The Parsing logic is
still correct according to the data returned in the latest (226) docs.

Fixes #285
[#95778668]