Rework download application bits
This commit reworks the test for all the download operations so they don't
have to assemble two complete byte arrays to compare.

It also parametrises the internal buffering, so that tests can run with
artificially small internal buffers, and fixes a bug in the implementation of
getStream whereby it calls subscriber.onNext() beyond the throttling of the
next stage, which may be take(), for example.

Finally, extra download tests are added with artificially reduced internal
buffers, so as to drive internal buffering code.

[#107069258]