Compiler Enforcement of Warnings
Previously, when the code had warnings in it, they were logged, but
since no one looks at the build logs, they were missed and continued
to live on in the code.  This change updates the compiler settings to
enforce that if there are warnings, the build fails.  It also updates
the code to remove a number of warnings that had been missed to this
point.

[#109949350]