Test Improvements
This change updates the tests in the -client-spring and -operations
projects.  The goal is to encapsulate the fact that each of the APIs
requires three tests (occasionally two) and that some of the setup
work is duplicated between the three test methods.

This work also identified some problems in how targeted organizations
and spaces are acquired.  Those problems were fixed as well.