TestObjects class for generating test objects.
TestObjects is a helper class for populating builder objects with test values
using a fill(BuilderType, String) method. Object setters are called with test
values based upon the setter parameter type. Setters which are collections or
belong to @Singular versions of collections are ignored.

String types are set to "test-"+modifier+settername, where <modifier> is the
second parameter to fill(b,modifier), and settername is the name of the
specific setter.

Boolean types are set to true.

Integer types are set to 1.

Types ending in Entity or Metadata are recursively filled, with the same
modifier, if their builder types can be found.

Paginated builder objects are treated specially, to set page request fields
consistently with the operations implementations.

This commit includes a TestObjectsTest class fills CreateApplicationRequest,
ListApplicationsRequest and ListApplicationsResponse with explicit checks.
It also adds a multi-page test to DefaultSpaceQuotasTest.list, just
to see how hard it is, and transformed versions of DefaultSpacesTest.java and
DefaultRoutesTest.java, to validate the approach.

Javadoc is included.

[#110330194]