Create a Domain owned by the given Organization (v2)
This change adds the v2 API for the deprecated Create a Domain owned by
the given Organization (POST /v2/domains).

This is the same as the Create a Shared Domain operation, they are on
the same endpoint and take and return the same data. I have merged
them in to a single createDomain operation.

[#110001838]