TestSubscriber Improvements
This change makes some improvements to the TestSubscriber to make it
fail in the appropriate way.  In addition, changes were made to tests
where the TestSubscriber wasn't being used appropriately (not calling
verify() mostly).  Note that this work has highlighted an issue in
Reactor that is causing a number of tests to fail.