Polishing
The code in #335 illustrated that we were treating certain response 
types (EventResource and OrganizationResource) differently than the 
rest.  This change updates the code to handle those the same as the 
other response types.  It also does some general formatting and code 
polishing.  Finally, it removes two response types that were never 
actually returned.