Fix delombok problem with Filters
Previously the delombok plugin had problems with method annotations that had
more than one argument.  There seems to be no way around this (something about
a bug in the compiler, see [1]) so another strategy to communicate the same
information needed to be devised.  In this case, the solution was to use an
@FilterParameter annotation placed on annotations for each of the types of
Filters.

[1]: https://github.com/rzwitserloot/lombok/issues/735