Supported API Version
Previously, the supported API version could eventually become greater than the
version of the Cloud Foundry instance being tested against.  In this case, the
InfoTest would fail. At the moment, I don't believe that this is so much a
failure as us being a bit ahead of the curve.  This change updates the test to
ensure that the version is greater than or equal to the Cloud Foundry instance
being tested against.