Operations Structure
This change adds some package structure to the operations project. Previously,
the entire content of the code was in (more or less) a single package.  As the
implementation has grown, and the trend is towards quite a large number of
classes, it's time to break things up.  The packages are scoped to the top-
level interfaces.