Map route operation
Stylistic note: passing along the validated request using `then` would
introduce an asymmetry between the signatures of the subsequent requestXxx
functions. It is convenient for all except the next one to be passed the
relevant request contents as parameters, so we do this for the first one also
and use `after` to discard the validated request.

[#106155542]