Strongly Typed Tuples
Previously, all uses of Tuple were immediately dereferenced in order to
maintain readability.  This was menial and repetitive and only required
because Java does not have a first-class Tuple type.  This change adds
factory methods that create Consumers, Functions, and Predicates which
dereference Tuples and pass them as arguments to inner functions.  This
improves readability and maintainability.

Note that the implementation of Tuples and its associated interfaces are here
temporarily until they are included in Reactor.