Improve cleanup
Previously there were two different locations where CloudFoundry cleanup was
done. The first was in the IntegrationTestConfiguration which attempted to
delete absolutely everything, but only considered organizations and spaces.
The other was in the @Before method of AbstractIntegrationTest and this
cleaned up everything, but filtered out the default organization and default
space. This change unifies these cleanups into a single utility which is
called with varying filters, but ensures that all cleanup is done
consistently.