Enforce JRE compliance
This change updates the maven build to include the animal-sniffer plugin. This
plugin enforces that most of the projects adhere to only Java 1.7 APIs even in
the presence of a Java 8 JDK.  It also updates the compiler in IntelliJ to
enforce the same failures as the compiler at the command line.