Ignore Integration Test
SpacesTest.listFilterByDeveloperId only works if the (global) user used by the
tests is not automatically a DeveloperId for a (global) space. This is true
for our local IntegrationTest environment, but not for CI IntegrationTest.

The proper solution is to create a new User for the tests, but this is not yet
available, so this test is ignored with a pointer to the story that provides
that function: https://www.pivotaltracker.com/story/show/101522686 — this is
consistent with other tests that need a fresh User.