Rework the scale operation.
Scale a named application in the targeted space by optionally adjusting the
memory limit, the disk limit and/or number of instances. If app is STARTED
then restart it after memory limit or disk limit changes, but NOT if the only
change is to instances.

If there are no changes the operation is a no-op.

This commit simplifies the top-level implementation, adds missing validation
logic, adds the restart functionality and simplifies stage processing. It also
adds a validation test and a memoryLimit/diskLimit parsing step.

Reformatting source code occurred in this commit.

[#106155436]