Add the get application events operation.
The CLI `cf events APP` by default gets the most recent 50 events for the
application. This implementation adds the maxNumberOfEvents() optional
parameter which permits adjusting this figure.

Some refactoring of DefaultApplicationTest occurred in this commit and
`events` was added to AbstractOperationsTest for the first time.

GetApplicationRequestTest was added and RestartApplicationRequestTest,
StartApplicationRequestTest, StopApplicationRequestTest and
RenameOrganizationRequestTest were modified to remove the use of fill() when
testing request validity.

[#106155452]