Redesign the Clients
Previously the SpringCloudFoundryClient and SpringLoggingClient were quite
complex and intertwined.  This made testing them in isolation very difficult.
This change updates the design to be much more decoupled and testable.  In
addition much of the complexity that used to be contained within the
SpringCloudFoundryClient has been broken out into smaller composable types
that are easier to test in isolation.

Finally, this change also changes the way that SSL certificate testing is
handled.  Instead of writing a truststore to the filesystem and hooking into
the default Java networking configuration, it keeps the SSL certificate
truststore in memory allowing it to be updated at any time in the startup
lifecycle.  This is intended to work around the fact that the Cloud
Controller, UAA, Logging, and Router all use different hosts and therefore
they may possibly use different certificates.