Change .after(supplier) calls to as(afterComplete(supplier))
Previously we were using .after() for the cases when we wished a result (or
the absence of a result) to be ignored, but _not_ an error.  This is not the
behaviour of .after(Supplier): this calls the Supplier on Error AND on
Complete, runs the resulting Mono and then passes complete, or the error(!),
downstream.

Instead of this behaviour we want the Supplier to be called only if the source
stream(Mono) completes without an error, and the resulting Mono to be run only
in this case.  Errors upstream should be passed on immediately without calling
the Supplier.

This commit adds to OperationUtils two methods afterStreamComplete() and
afterComplete() and between them they fill this gap.  24 places in the
production code and the tests are updated to use them. Some of these usage
points were simplified.

[#113814653]