Fix RoutesTest#createInvalidDomain test
Previously the test was detecting a failure in createDomain, not in create
Route.  This commit puts a createDomain test in the DomainsTest integration
test class (which checks the right exception) and reworks the
createInvalidDomain test to become createRouteWithNonExistentDomain, which
better reflects the CloudFoundryException actually returned by the
routes().create() operation. Both tests are simpler than the original.

[#113929343]