JWT Username
Previously, the username of the current user was fixed to the username
configured when the client was created.  This will not be dynamic enough for
when multiple threads will be using different identities.  This change
refactors the username determination to decode the username from the JWT token
being used for authentication.

In order to do this, a UaaClient needed to be created so that the JWT signing
key could be retrieved.  The introduction of this client illustrated that the
code was packaged up improperly.  This change also splits the
CloudFoundryClient, the LoggingClient, and the UaaClient into different
package trees separated from one another.  This also preps the project for the
eventual introduction of a RouterClient.