Speed Up Integration Tests
Previously, the integration tests took nearly 20 minutes to run.  Facing a
massive expansion of the number of integration tests, we had to speed up the
testing.  It was identified that the biggest slowdown in the tests was the
async cleanup of entities created during the test run.  The biggest win would
be to parallelize this cleanup, but that meant that the tests had to cope with
a "dirty" system during the run themselves.  This meant that every test has to
use unique names for their entities and filter more stringently than they did
before.

[#113553669]