Add push operation
This commit adds the push operation, with the exception of the manifest
option.

In this commit there are also some changes to the behaviour of restart() and
start(), so that staging is explicitly tested for completion before waiting
for the instances to be running. The corresponding unit tests are adjusted to
match.

[#106155434]