Rewrite spaces().get() to avoid bug and correct bad returns
spaces().get() now generates rules when --securityGroupRules flag is set, and
doesn't if it is omitted, or false. The SpaceDetail object is expanded to
include a list of rules with each securityGroup it reports; the list is empty
if the flag is not set.

Optional spaceQuota is retained, but is only empty if there is no spaceQuota
set in the space.

A test has been added and some tests modified to check these additions.

[#114402261]