IdentityZone Types
This change rationalizes the identity zone types.  Specifically, it creates an
AbstractIdentityZone that all related types extend from.  The list identity
zone command then returns a concrete type of IdentityZone.  In addition, the
ListIdentityZoneResponse type now is a List to properly match the return type.
I'm not completely committed to this, but I think it's the right choice at the
moment.

[#114245183][resolves #410]