Disable Incorrect Inspection
IntelliJ has an inspection that verifies that @Autowired is only used in beans
that are valid Spring Beans, i.e. @Component, @Service, etc.  The problem is
that this inspection is incorrect as it ignores the fact that @Autowired can
be used in test classes.  This discrepancy caused all of our integration
tests, which use `@Autowired` greatly, to show warnings.  This change disables
that inspection (we're pretty good Spring developers) to remove the warnings.