Add create service instance operation
This commit adds the operation create service (instance) and tests. It also
creates AbstractServiceInstanceResource and adds Lombok @Data to requests in
the serviceinstances package which had omitted it. The create operation waits
for creation to finish before completion indicated by the state of the last
operation on the service instance.

[#106155480]