Add deleteWithServiceBindings client Applications integration test
This commit adds a test which binds a service instance to an application
before deleting it. The service instance is a user-provided one to avoid
assuming the existence of available services in the test environment.

In order to get this to work, a "union type" subsuming both managed and user-
provided services was created. The definitions of
AbstractServiceInstanceResource and ServiceInstance{Resource, Entity} were
changed to use a BaseServiceInstanceEntity (containing the common elements to
both), and UnionServiceInstance{Resource, Entity} as well as
AbstractUnionServiceInstanceResource were created. These are used in the
responses to general list service instances request responses, for example
ListSpaceServiceInstanceResponse.

[#115938635]