Fix NPE in delete application operation
Previously, if deleteRoutes() was omitted from the DeleteApplicationRequest,
then the operation implementation would throw a NullPointerException. This
commit implements a default of false (do not delete routes), adds unit tests
to check we observe it, and modifies integration tests to use the default when
appropriate.

Note: aggressive use of the fill() test utility masked this bug during
development.

[#116597965]