Short Circuit Completed Jobs
Previously, when a job entity was returned from an async REST call, the
JobUtils.waitForCompletion() would always result in making at least one
additional network call.  This change inspects that first returned result to
determine if it is finished and if it is, short circuits the rest of the
behavior resulting on zero further network calls.

[#115168301]