Last operation missing from ServiceInstance output
This commit adds an extra field to the ServiceInstance output on operations
getInstance and listInstances.

In the CLI the last operation type and status are returned in a composite
field called "status". In our more program-friendly interface we provide
"message", "status" and "lastOperation" for the ServiceInstance taken
respectively from the "message", "status" and "type" fields of the
LastOperation information available from the client API.

[#106155478]