Multi-page Compensation
Previously, paginated requests were consumed as soon as they arrived.  Since
it was possible that the results were used in a way that modified the list, it
was possible that the complete collection of the results was not consistent.
To protect against this problem, the results of the pagination are now held
until the entirety is enumerated.  The enumeration still happens in parallel
and the results can be consumed in parallel, but there is an explicit barrier
to prevent consistency problems.