Code Relocation
The original contribution of this code followed all of our design heuristics
to the letter.  However, the design of this particular set of documentation is
a bit too detailed and doesn't group the API logically with others like it.
This change updates the code, primarily by moving and renaming to more
accurately represent where it should logically exist.

[#101522638][resolves #452]