Filters Discard Empty Values
A previous change updated the FilterBuilder to remove null values from
collections before processing them.  This change helped a few spurious errors,
but didn't take into account the possibility of an "empty" value as well.  In
general we wouldn't filter for a value like this, but certain CC APIs return
empty Strings when they should omit the key completely.  This change causes
the FilterBuilder to filter out "empty" values as well.