Fix Error When Binding Bound App
Previously the Operations API would return an error if the user tried to
bind an application to a service it was already bound to. In the same
scenario the CLI will return success.

This change swallows the specific error CF produces in this scenario, so
the user sees a successful (no-op) bind.

[#118128115]