Polishing
References to lambda function parameters called `request1` are changed to
`validRequest` everywhere. This affected about 100 uses, all of which were to
avoid re-using the (method) parameter name `request` after it (the request)
had been validated prior to use in the implementation. The new name is
considered to be more accurate and harder to mis-type which might result in an
obscure bug.