Polishing
This change updates the contribution by removing a name filter parameter.
While the original PR had this, there's no indication that it still exists in
the mainline code, and the documentation does not list it.  As it is not
required for the existing functionality (of both the CLI and the operations
layer), it has been removed.

[#118857947][resolves #475]