Completing List Service Brokers Operation
This change finishes the implementation of the List Service Brokers Operation
by completing a couple of missing pieces.  First, an id was added to the
Server Broker type to enable easy fallback to the Client API.  Second, the
ServiceAdmin type was exposed via CloudFoundryOperations interface allowing
users to get access to the operation.  Finally there was a bit of polishing
around other bits of the code.

[#106155624][resolves #481]