Immutables
Over time, our choice of Lombok has proven to be troublesome.  There are
problems with how it operates in our build lifecycle, the requirements for
IDEs, Javadoc verbosity, and functional problems about initializing empty
collections.

This change is the first step towards another project called Immutables.  This
is a tentative conversion and will be used to evaluate whether we should make
the jump completely to this project.