Use Immutables to send empty collection payloads
Previously, our use of Lombok made it so that we could not send empty, but
non-null collection payloads.  This caused issues with things like emptying
the collection of environment variables when updating an application.  This
change switches the UpdateApplicationRequest to use Immutables which should
solve this issue.

This is more testing towards converting the entire library to use Immutables
instead of Lombok.

[#116804259]