Augment TestSubscriber to detect failure to subscribe
In integration tests, TestSubscriber verification happens in generic @After
code, which waits on a private latch for the source stream to complete (or
error).  If, for any reason, the TestSubscriber never actually subscribes to
the main stream this latch never falls and the @After code waits (for 5
minutes) before failing.

Immutables-generated requests throw validation exceptions on build(), which
may result in unnecessary 5-minute delays.

This commit adds a simple boolean flag to detect if onSubscribe() has been
called on the TestSubscriber. The verification code then throws an immediate
exception if it has not.  This avoids waiting for a completion that will never
happen.