Improvements to UAA List Users
This change makes some improvements to the implementation of List Users and
the supported pagination code added for it.  First, the Pagination and Sorting
types were combined into one because there is no indication, in the
documentation or the UAA source code that they could ever exist apart. Second,
the SortOrder type was moved out of the  users package to be a peer of the
PaginatedRequest.  Third, The ListUsersResponse type had the incorrect
getUsersMethod removed.  Finally, some improvements were made to the
formatting of the various types that were added.

[#114246743]