Allow Map Route operation to leave path unspecified
Previously, a null path resulted in errors in the implementation of the map-
route operation. This commit protects the implementation steps from a null
path, and extends the integration tests to check that null paths are allowed
on map and unmap and that they correctly add and remove the mappings so
produced.

[#120802459]