CloudFoundryOperations Builder
Previously, the CloudFoundryOperationsBuilder used a builder pattern that was
unlike any other builder in the system.  Arguably it was a bit nicer than the
others, but the lack of consistency lead to confusion for users.  This change
makes the builder consistent with the others in the system, and updates the
documentation to match.  In addition, it takes into account the possibility
that the DefaultCloudFoundryOperations is not configured with a
CloudFoundryClient (Advanced#sshCode is an example of where this is warranted).