UAA Users Integration Tests
This change adds integration tests for the UAA Users functionality, and handles
some edge cases in the UAA api such as the default admin user not having a
family name or given name.

It introduces a new environment variable, test.protected.users, which *must* be
set for tests to run.

[#114247177]