Augment tests for empty collections on update requests
Previously, with the Lombok implementation of request classes, optional
collections were omitted if empty. In the case of Update*Requests sometimes
an empty collection meant just that, and omitting it failed to update the
collection.

With the move to Immutables, this problem appeared to be improved (@Nullable
collections can be null). Although a specific bug was fixed in the operations
integration tests for UpdateApplication, a number of other Update requests
were not checked. This commit adds or augments tests for
UpdateApplicationRequest, UpdateRunningEnvironmentVariablesRequest,
UpdateStagingEnvironmentVariablesRequest, UpdateServiceInstanceRequest,
UpdateSpaceRequest and UpdateUserProvidedServiceInstanceRequest.

Unit tests were added to check that an empty collection generated the correct
URI request. Client API Integration Tests were added to check that CF
correctly interpreted our representation.

[#120266741]