First use of Optionals as API
Since the conversion from Lombok to Immutables, we've been struggling with
Optionality/Nullability.  Lombok took a default "anything can be null" while
Immutables takes the "you must explicitly make things null" position.  We're
fans of the latter, but it works in opposition to Postel's Law and the client
being liberal in what it accepts.  For the last couple of weeks we've been
finding (because there's no way to know for sure from the documentation) that
there are properties in the CF API that can be null and we hadn't expected it.

We know that we need to be more liberal in what we accept certainly, and to
date, we've done this by using an @Nullable parameter.  If we are looking to
expand this nullability however, we're very concerned about parts of the API
becoming effectively "null anywhere", but not indicating that to the users via
signatures.

This is the first commit as we convert from using @Nullable to using
Optionals.  Optionals have downsides in that they are more difficult to use
and generally require an extra dereference, but we believe that the benefit of
becoming explicit about where we thing values can be optional are worth it.
This is not the final judgement however, and if users chafe under this change
it may come back.