All CF responses nullable
This change updates all of the response types in the CF API so that all
properties are nullable.  Since the current API documentation does not
indicate what is nullable and what is not, this is the safest way to handle
the ambiguity.  Unlike the earlier attempt at using Optionals, this change
requires all callers to be aware that nulls may be returned.