Fix CI Failure
An earlier commit changed the annotation on the _ServiceInstance Tags field to
be nullable. This clashed with an earlier change to the Fill function, which
meant that Tags was now being populated (with an empty list) rather than being
nulled. This change modifies the DefaultServicesTest to explicit null the Tags
field, thereby avoiding the fill value and once again meeting the test
expectation.