Authorization Code Grant (Browser) doesn't return code
Previously it was thought that this API call (in uaa.authorizations) should
return an authorization code (as documented) but instead it returns a redirect
location URI for an authentication protocol/page.  This commit updates the
return value of this request to reflect this, and updates unit and integration
tests accordingly.

[#120471467]