Polishing
Previously clientId and clientSecret were beans created in integration test
only used in uaa/TokensTest. This commit renames these to testClientId and
testClientSecret to avoid confusion with the standard id and secret used for
creating the cloudfoundry client.  Comments are added to the bean definitions
to indicate where they are used (which is otherwise obscured by Spring's bean
injection).