Doppler Envelopes
Previously the implementation of the Doppler APIs stripped the envelopes from
the response values, making the responses as specific to the requests as
possible.  This was done because the act of getting the "real" payload out was
a bit of a pain (inspecting a value, using that result to get another value
specific to it).  The assumption was that nothing in the envelope was
important other than the actual event.  It turns out that this assumption was
correct.  An issue was opened demonstrating that the contents of the envelope
were useful and therefore shouldn't be discarded.

This change updates the Doppler APIs to return the envelopes for all events.
The caller of the API is now responsible for inspecting the type and
extracting the "real" event accordingly.  In addition the two places we use
the Doppler API (application operations around logging) have been updated to
do this extraction.

[#120697805][resolves #489]