Builder API CHanges
This change, as massive as it seems was intended to change to the API of the
*Client builders.  In addition to that, it had the side effects of removing
all of the rest of the Lombok usage, and even removal of Spring for all places
where we access the network.  There are still some places where we use Spring
utilities (UriComponentBuilder mostly), but the vast majority of it is gone.

This is probably the biggest change on our way to API stability, so it's great
to finally get it out of the way.