Cleanup cleanup
Previously, the Cloud Foundry Cleaner was very destructive and required
declarations of the things that it needed to preserve.  This change updates
the cleaner to only cleanup items created by the tests.  It determines which
items these are by filtering against the naming patterns for all items as
created by the AbstractIntegrationTest.