Add tags in service offering class
Currently, the controller returns a list of tags for each service offering
exposed in the /services endpoint. These tags, however, are ignored by the
Java client and are not part of the ServiceOffering class that is returned
to the user.

These tags can be useful to tools that automate the management of services
(to determine whether a service's tags should be updated  or not). Because
of this, this commit adds a getTags( ) method in the ServiceOffering class
that allows users to access them.