InputStreams in Operations API changed to Paths
Previously, the Operations layer had two places where InputStreams were used
to push binary data to the server.  In one of those places, that binary data
needed to be examined so that resource matching could reduce that amount of
data sent over the wire, speeding up transfers.  Unfortunately, without
jumping through many hoops, it wasn't going to be possible to inspect the
contents of the InputStream.

This change updates those two places in the Operations layer to use
java.nio.file.Paths instead.  Using these Paths, we can examine a view of the
filesystem, exploded, within an archive, or any other java.nio.file.FileSystem
implementation currently accessible to the application.  While we don't use
this functionality today (we simply form an InputStream for delegation to the
Client layer), we now have a stable API to implement the resource matching and
upload slimming behavior with breaking backwards compatibility.