Fix Create/Update client requests' List parameters
Previously, some list parameters on _*Request classes were not marked
@Nullable, since they were required. However, not setting them did not provoke
an error but instead produced an empty list value. This is not the correct
default behaviour. This commit sets them to @Nullable, which leaves them null
if not set explicitly, and also adds a @Value.Check to the Request class to
explicitly check for null, so as to throw the correct IllegalStateException if
not set.  This combination produces the desired behaviour.

For consistency, the text of each explicit IllegalStateException thrown from a
Check method is updated to be similar to that issued by Immutables'
boilerplate code.

[#126488735][resolves #526]