Token Refresh
Previously, the implementation of the PasswordGrantTokenProvider only
requested a single token.  This caused issues if the TokenProvider was kept
around longer than a token's expiration (5 minutes).  This change adds repeat
step that refreshes the token before its expiration.  It also adds an abstract
class that encapsulates the refresh behavior for so that implementations can
focus on retrieving the initial access token.